Fixed the build, organized the project in files
