+++ .vscode/settings.json
@@ -1,5 +1,8 @@
 {
     "files.associations": {
-        "stdlib.h": "c"
+        "stdlib.h": "c",
+        "powerconverter.h": "c",
+        "temperatureconverter.h": "c",
+        "program.h": "c"
     }
 }
