+++ .vscode/tasks.json
@@ -25,6 +25,30 @@
             "problemMatcher":[
                 "$gcc"
             ]
+        },
+        {
+            "label": "Compile",
+            "type": "cppbuild",
+            "command": "C:/TDM-GCC-64/bin/gcc.exe",
+            "args":[
+                "-Wall",
+                "-I",
+                "${fileDirname}/Libs",
+                "${fileDirname}/Converters/*.c",
+                "${fileDirname}/*.c",
+                "-o",
+                "${fileDirname}/build/program.exe"
+            ],
+            "options": {
+                "cwd": "${fileDirname}"
+            },
+            "group":{
+                "kind": "build",
+                "isDefault": true
+            },
+            "problemMatcher":[
+                "$gcc"
+            ]
         }
     ]
 }
