+++ Converters/powerConverter.c
@@ -1,4 +1,4 @@
-#include "PowerConverter.h"
+#include "../Libs/powerConverter.h"
 #include <stdio.h>
 
 // Implementação das funções de conversão
