+++ Converters/temperatureConverter.c
@@ -1,4 +1,4 @@
-#include "temperaturaconverter.h"
+#include "../Libs/temperatureConverter.h"
 
 double CelsiusToFahrenheit(double celsius) {
     return (celsius * 9.0 / 5.0) + 32.0;
