+++ Libs/temperatureConverter.h
@@ -1,5 +1,5 @@
-#ifndef TEMPERATURACONVERTER_H
-#define TEMPERATURACONVERTER_H
+#ifndef TEMPERATURECONVERTER_H
+#define TEMPERATURECONVERTER_H
 
 double CelsiusToFahrenheit(double celsius);
 double CelsiusToKelvin(double celsius);
