+++ README.md
@@ -10,9 +10,9 @@
 
 #### Instruções de compilação
 Para compilar:
-* 1.Navegue até a pasta do projeto no terminal;
-* 2.Digite gcc *.c -o output/program.exe
-* 3.Depois execute com .\output\program.exe
+* 1.Abra o programa main.c ou program.c do VS CODE;
+* 2.Aperte "CTRL + SHIFT + B" e escolha a opção "Compile" para compilar um novo executável;
+* 3.Depois execute com .\build\program.exe
 
 ### Instruções de depuração
 Para debugar:
