+++ main.c
@@ -1,8 +1,8 @@
 #include<stdlib.h>
-#include "program.h"
+#include <program.h>
 
 int main(){
-    system("CHCP 65001>consoleConfig"); //Configura a coleção de caracteres do console e salva em consoleConfig
+    //system("CHCP 65001>consoleConfig"); //Configura a coleção de caracteres do console e salva em consoleConfig
     GetHeader(); //Chama o cabeçalho
     GetMenu(); //Chama o menu
     GetFooter(); //Chama o rodapé
